fix upgrade bug
